Moved styles to separate .js file for modularity
